extract firebase from utils folder
